requirements.txt
databricks-vectorsearch
streamlit
langchain
langchain-community
openai
langchain-openai